+++ lib/rib/core/strip_backtrace.rb
@@ -7,7 +7,6 @@ module Rib::StripBacktrace
 
   # --------------- Rib API ---------------
 
-  # strip backtrace until rib
   def format_error err
     return super if StripBacktrace.disabled?
     message, backtrace = get_error(err)